strncmp perfect paco aproved

+++ libft.h
@@ -21,7 +21,7 @@ size_t	ft_strlen(const char *s);
 char	*ft_strchr(const char *s, int c);
 char	*ft_strrchr(const char *s, int c);
 char	*ft_strdup(const char *s);
-int		ft_strncmp(char *s1, char *s2, unsigned int n);
+int     ft_strncmp(char *s1, char *s2, size_t n);
 int		ft_isdigit(int c);
 int		ft_isalpha(int c);
 int		ft_isalnum(int c);
